server/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-engineio==3.2.3
python-socketio==3.1.1
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
